+++ src/components/ProfileContainer.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import ProfileBox from './ProfileBox'
+const ProfileContainer = ({ profiles }) => {
+  return (
+    <>
+      {profiles.map((p) => (
+        <ProfileBox key={p.profileId} profile={p.profile} />
+      ))}
+    </>
+  )
+}
+
+export default ProfileContainer
